requirements.txt
pandas
numpy
faker
matplotlib
seaborn
paho-mqtt
sqlite3
kaggle
